# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YoYo Map API — global yo-yo player community map
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ApiErrorError = {
  code: string;
  message: string;
  requestId?: string;
};

export interface ApiError {
  error: ApiErrorError;
}

export type MapEntryEntityType =
  (typeof MapEntryEntityType)[keyof typeof MapEntryEntityType];

export const MapEntryEntityType = {
  person: "person",
  shop: "shop",
  club: "club",
} as const;

export interface MapEntry {
  id: string;
  display_name: string;
  city: string;
  region?: string | null;
  country: string;
  lat: number;
  lng: number;
  entity_type?: MapEntryEntityType;
  verified_owner?: boolean | null;
}

export type MapEntryDetailSocials = { [key: string]: string };

export interface MapEntryDetail {
  bio?: string | null;
  socials: MapEntryDetailSocials;
  address_line?: string | null;
  postal_code?: string | null;
  hours?: string | null;
  club_meeting_info?: string | null;
  club_venue_public?: boolean | null;
}

export interface EntryCounts {
  person: number;
  shop: number;
  club: number;
}

export interface Country {
  id: number;
  code: string;
  name: string;
}

export interface Region {
  id: number;
  name: string;
  country_id: number;
}

export interface City {
  id: number;
  name: string;
  country_id: number;
  region_id?: number | null;
}

export interface LocationsResponse {
  countries?: Country[];
  regions?: Region[];
  cities?: City[];
}

export interface CreateCityBody {
  name: string;
  countryId: number;
  regionId?: number | null;
}

export type SubmitBodyEntityType =
  (typeof SubmitBodyEntityType)[keyof typeof SubmitBodyEntityType];

export const SubmitBodyEntityType = {
  person: "person",
  shop: "shop",
  club: "club",
} as const;

export type SubmitBodyAgeBand =
  (typeof SubmitBodyAgeBand)[keyof typeof SubmitBodyAgeBand];

export const SubmitBodyAgeBand = {
  "13-17": "13-17",
  "18+": "18+",
} as const;

export type SubmitBodySocials = { [key: string]: string };

export interface SubmitBody {
  entityType: SubmitBodyEntityType;
  displayName: string;
  email: string;
  city_id?: number | null;
  region_id?: number | null;
  country_id?: number | null;
  bio?: string;
  ageBand?: SubmitBodyAgeBand;
  parentName?: string;
  parentEmail?: string;
  relationship?: string;
  addressLine?: string;
  postalCode?: string;
  hours?: string;
  contactName?: string;
  authorizedRep?: boolean;
  clubMeetingInfo?: string;
  clubVenuePublic?: boolean;
  venueAddressLine?: string;
  venuePostalCode?: string;
  socials?: SubmitBodySocials;
  consentPrivacy: boolean;
  consentTerms: boolean;
  consentPublic: boolean;
  honeypot?: string;
}

export type SubmitResultEmailStatus =
  (typeof SubmitResultEmailStatus)[keyof typeof SubmitResultEmailStatus];

export const SubmitResultEmailStatus = {
  sent: "sent",
  queued: "queued",
  failed: "failed",
  partial_failed: "partial_failed",
} as const;

export interface SubmitResult {
  message: string;
  emailStatus?: SubmitResultEmailStatus;
  requestId?: string;
  retryAt?: string | null;
  isMinor?: boolean;
}

export type ProfileEntrySocials = { [key: string]: string };

export type ProfileEntryEntityType =
  (typeof ProfileEntryEntityType)[keyof typeof ProfileEntryEntityType];

export const ProfileEntryEntityType = {
  person: "person",
  shop: "shop",
  club: "club",
} as const;

export interface ProfileEntry {
  id: string;
  display_name: string;
  city: string;
  region?: string | null;
  country: string;
  bio?: string | null;
  socials: ProfileEntrySocials;
  is_visible: boolean;
  entity_type: ProfileEntryEntityType;
  club_meeting_info?: string | null;
  club_venue_public: boolean;
  contact_name?: string | null;
  address_line?: string | null;
  postal_code?: string | null;
  hours?: string | null;
}

export type UpdateProfileBodySocials = { [key: string]: string };

export interface UpdateProfileBody {
  token: string;
  display_name?: string;
  bio?: string | null;
  socials?: UpdateProfileBodySocials;
  club_meeting_info?: string | null;
  club_venue_public?: boolean;
  contact_name?: string | null;
  address_line?: string | null;
  postal_code?: string | null;
  hours?: string | null;
}

export type ReportBodyReason =
  (typeof ReportBodyReason)[keyof typeof ReportBodyReason];

export const ReportBodyReason = {
  spam: "spam",
  harassment: "harassment",
  impersonation: "impersonation",
  minor_unsafe: "minor_unsafe",
  other: "other",
} as const;

export interface ReportBody {
  entryId: string;
  reason: ReportBodyReason;
  details?: string;
  reporterEmail?: string;
}

export type ListMapEntries200 = {
  entries: MapEntry[];
  counts: EntryCounts;
};

export type ListLocationsParams = {
  type: ListLocationsType;
  countryId?: number;
  regionId?: number;
};

export type ListLocationsType =
  (typeof ListLocationsType)[keyof typeof ListLocationsType];

export const ListLocationsType = {
  countries: "countries",
  regions: "regions",
  cities: "cities",
} as const;

export type CreateCity200 = {
  city: City;
};

export type SendMagicLinkBody = {
  email: string;
};

export type SendMagicLink200 = {
  message: string;
};

export type VerifyMagicLinkParams = {
  token: string;
};

export type VerifyMagicLink200 = {
  entry: ProfileEntry;
};

export type UpdateProfile200 = {
  message: string;
};

export type DeleteProfileBody = {
  token: string;
};

export type DeleteProfile200 = {
  message: string;
};

export type ReportEntry200 = {
  message: string;
};