# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: YoYo Map API — global yo-yo player community map
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /entry/{id}:
    get:
      operationId: getEntryDetail
      tags: [map]
      summary: Get full entry detail for popup
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Entry detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapEntryDetail"
        "404":
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /locations:
    get:
      operationId: listLocations
      tags: [locations]
      summary: List countries, regions, or cities
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [countries, regions, cities]
        - name: countryId
          in: query
          schema:
            type: integer
        - name: regionId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Location data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationsResponse"
    post:
      operationId: createCity
      tags: [locations]
      summary: Add a new city
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCityBody"
      responses:
        "200":
          description: Created city
          content:
            application/json:
              schema:
                type: object
                properties:
                  city:
                    $ref: "#/components/schemas/City"
                required: [city]
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /auth/magic-link:
    post:
      operationId: sendMagicLink
      tags: [auth]
      summary: Send magic link email for profile management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required: [email]
      responses:
        "200":
          description: Magic link sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required: [message]
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: No entry found for that email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/verify-link:
    get:
      operationId: verifyMagicLink
      tags: [auth]
      summary: Verify magic link token and return entry data
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Token verified, returns entry
          content:
            application/json:
              schema:
                type: object
                properties:
                  entry:
                    $ref: "#/components/schemas/ProfileEntry"
                required: [entry]
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
          required: [code, message]
      required: [error]
...
    MapEntryDetail:
      type: object
      properties:
        bio:
          type: string
          nullable: true
        socials:
          type: object
          additionalProperties:
            type: string
        address_line:
          type: string
          nullable: true
        postal_code:
          type: string
          nullable: true
        hours:
          type: string
          nullable: true
        club_meeting_info:
          type: string
          nullable: true
        club_venue_public:
          type: boolean
          nullable: true
      required: [socials]
...
    Country:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
      required: [id, code, name]
    Region:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        country_id:
          type: integer
      required: [id, name, country_id]
    City:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        country_id:
          type: integer
        region_id:
          type: integer
          nullable: true
      required: [id, name, country_id]
    LocationsResponse:
      type: object
      properties:
        countries:
          type: array
          items:
            $ref: "#/components/schemas/Country"
        regions:
          type: array
          items:
            $ref: "#/components/schemas/Region"
        cities:
          type: array
          items:
            $ref: "#/components/schemas/City"
    CreateCityBody:
      type: object
      properties:
        name:
          type: string
        countryId:
          type: integer
        regionId:
          type: integer
          nullable: true
      required: [name, countryId]
...
    ProfileEntry:
      type: object
      properties:
        id:
          type: string
        display_name:
          type: string
        city:
          type: string
        region:
          type: string
          nullable: true
        country:
          type: string
        bio:
          type: string
          nullable: true
        socials:
          type: object
          additionalProperties:
            type: string
        is_visible:
          type: boolean
        entity_type:
          type: string
          enum: [person, shop, club]
        club_meeting_info:
          type: string
          nullable: true
        club_venue_public:
          type: boolean
        contact_name:
          type: string
          nullable: true
        address_line:
          type: string
          nullable: true
        postal_code:
          type: string
          nullable: true
        hours:
          type: string
          nullable: true
      required: [id, display_name, city, country, socials, is_visible, entity_type, club_venue_public]
...