# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YoYo Map API — global yo-yo player community map
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all visible map entries (lean shape)
 */
export const ListMapEntriesResponse = zod.object({
  entries: zod.array(
    zod.object({
      id: zod.string(),
      display_name: zod.string(),
      city: zod.string(),
      region: zod.string().nullish(),
      country: zod.string(),
      lat: zod.number(),
      lng: zod.number(),
      entity_type: zod.enum(["person", "shop", "club"]).optional(),
      verified_owner: zod.boolean().nullish(),
    }),
  ),
  counts: zod.object({
    person: zod.number(),
    shop: zod.number(),
    club: zod.number(),
  }),
});

/**
 * @summary Get full entry detail for popup
 */
export const GetEntryDetailParams = zod.object({
  id: zod.coerce.string(),
});

export const GetEntryDetailResponse = zod.object({
  bio: zod.string().nullish(),
  socials: zod.record(zod.string(), zod.string()),
  address_line: zod.string().nullish(),
  postal_code: zod.string().nullish(),
  hours: zod.string().nullish(),
  club_meeting_info: zod.string().nullish(),
  club_venue_public: zod.boolean().nullish(),
});

/**
 * @summary List countries, regions, or cities
 */
export const ListLocationsQueryParams = zod.object({
  type: zod.enum(["countries", "regions", "cities"]),
  countryId: zod.coerce.number().optional(),
  regionId: zod.coerce.number().optional(),
});

export const ListLocationsResponse = zod.object({
  countries: zod
    .array(
      zod.object({
        id: zod.number(),
        code: zod.string(),
        name: zod.string(),
      }),
    )
    .optional(),
  regions: zod
    .array(
      zod.object({
        id: zod.number(),
        name: zod.string(),
        country_id: zod.number(),
      }),
    )
    .optional(),
  cities: zod
    .array(
      zod.object({
        id: zod.number(),
        name: zod.string(),
        country_id: zod.number(),
        region_id: zod.number().nullish(),
      }),
    )
    .optional(),
});

/**
 * @summary Add a new city
 */
export const CreateCityBody = zod.object({
  name: zod.string(),
  countryId: zod.number(),
  regionId: zod.number().nullish(),
});

export const CreateCityResponse = zod.object({
  city: zod.object({
    id: zod.number(),
    name: zod.string(),
    country_id: zod.number(),
    region_id: zod.number().nullish(),
  }),
});

/**
 * @summary Submit a new map entry
 */
export const SubmitEntryBody = zod.object({
  entityType: zod.enum(["person", "shop", "club"]),
  displayName: zod.string(),
  email: zod.string().email(),
  city_id: zod.number().nullish(),
  region_id: zod.number().nullish(),
  country_id: zod.number().nullish(),
  bio: zod.string().optional(),
  ageBand: zod.enum(["13-17", "18+"]).optional(),
  parentName: zod.string().optional(),
  parentEmail: zod.string().optional(),
  relationship: zod.string().optional(),
  addressLine: zod.string().optional(),
  postalCode: zod.string().optional(),
  hours: zod.string().optional(),
  contactName: zod.string().optional(),
  authorizedRep: zod.boolean().optional(),
  clubMeetingInfo: zod.string().optional(),
  clubVenuePublic: zod.boolean().optional(),
  venueAddressLine: zod.string().optional(),
  venuePostalCode: zod.string().optional(),
  socials: zod.record(zod.string(), zod.string()).optional(),
  consentPrivacy: zod.boolean(),
  consentTerms: zod.boolean(),
  consentPublic: zod.boolean(),
  honeypot: zod.string().optional(),
});

export const SubmitEntryResponse = zod.object({
  message: zod.string(),
  emailStatus: zod
    .enum(["sent", "queued", "failed", "partial_failed"])
    .optional(),
  requestId: zod.string().optional(),
  retryAt: zod.string().nullish(),
  isMinor: zod.boolean().optional(),
});

/**
 * @summary Send magic link email for profile management
 */
export const SendMagicLinkBody = zod.object({
  email: zod.string().email(),
});

export const SendMagicLinkResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Verify magic link token and return entry data
 */
export const VerifyMagicLinkQueryParams = zod.object({
  token: zod.coerce.string(),
});

export const VerifyMagicLinkResponse = zod.object({
  entry: zod.object({
    id: zod.string(),
    display_name: zod.string(),
    city: zod.string(),
    region: zod.string().nullish(),
    country: zod.string(),
    bio: zod.string().nullish(),
    socials: zod.record(zod.string(), zod.string()),
    is_visible: zod.boolean(),
    entity_type: zod.enum(["person", "shop", "club"]),
    club_meeting_info: zod.string().nullish(),
    club_venue_public: zod.boolean(),
    contact_name: zod.string().nullish(),
    address_line: zod.string().nullish(),
    postal_code: zod.string().nullish(),
    hours: zod.string().nullish(),
  }),
});

/**
 * @summary Update a profile entry
 */
export const UpdateProfileBody = zod.object({
  token: zod.string(),
  display_name: zod.string().optional(),
  bio: zod.string().nullish(),
  socials: zod.record(zod.string(), zod.string()).optional(),
  club_meeting_info: zod.string().nullish(),
  club_venue_public: zod.boolean().optional(),
  contact_name: zod.string().nullish(),
  address_line: zod.string().nullish(),
  postal_code: zod.string().nullish(),
  hours: zod.string().nullish(),
});

export const UpdateProfileResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Delete a profile entry
 */
export const DeleteProfileBody = zod.object({
  token: zod.string(),
});

export const DeleteProfileResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Report a map entry
 */
export const ReportEntryBody = zod.object({
  entryId: zod.string(),
  reason: zod.enum([
    "spam",
    "harassment",
    "impersonation",
    "minor_unsafe",
    "other",
  ]),
  details: zod.string().optional(),
  reporterEmail: zod.string().email().optional(),
});

export const ReportEntryResponse = zod.object({
  message: zod.string(),
});