# lib/api-zod/src/generated/types/apiError.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YoYo Map API — global yo-yo player community map
 * OpenAPI spec version: 0.1.0
 */
import type { ApiErrorError } from "./apiErrorError";

export interface ApiError {
  error: ApiErrorError;
}